+++ src/app/view/components/dot-toolbar/components/dot-toolbar-user/dot-toolbar-user.component.html
@@ -15,10 +15,10 @@
             </div>
         </header>
         <ul class="toolbar-user__menu" id="dot-toolbar-user-menu">
-            <li class="user-menu__item"><a href="#" (click)='toggleMyAccount()' *ngIf="!auth.loginAsUser" id="dot-toolbar-user-link-my-account" [textContent]="i18nMessages['my-account']"></a></li>
-            <li class="user-menu__item"><a href="#" (click)="tooggleLoginAs($event)" *ngIf="!auth.loginAsUser" id="dot-toolbar-user-link-login-as" [textContent]="i18nMessages['login-as']">login-as</a></li>
-            <li class="user-menu__item"><a href="#" (click)="logout($event)" *ngIf="!auth.loginAsUser" id="dot-toolbar-user-link-logout" [textContent]="i18nMessages['Logout']"></a></li>
-            <li class="user-menu__item"><a href="#" (click)="logoutAs($event)" *ngIf="auth.loginAsUser" id="dot-toolbar-user-link-logout-as" [textContent]="i18nMessages['logout-as']">logout-as</a></li>
+            <li class="user-menu__item"><a href="#" (click)='toggleMyAccount()' *ngIf="!auth.loginAsUser" id="dot-toolbar-user-link-my-account" [textContent]="'my-account' | dm "></a></li>
+            <li class="user-menu__item"><a href="#" (click)="tooggleLoginAs($event)" *ngIf="!auth.loginAsUser" id="dot-toolbar-user-link-login-as" [textContent]="'login-as' | dm ">login-as</a></li>
+            <li class="user-menu__item"><a href="#" (click)="logout($event)" *ngIf="!auth.loginAsUser" id="dot-toolbar-user-link-logout" [textContent]="'Logout' | dm "></a></li>
+            <li class="user-menu__item"><a href="#" (click)="logoutAs($event)" *ngIf="auth.loginAsUser" id="dot-toolbar-user-link-logout-as" [textContent]="'logout-as' | dm ">logout-as</a></li>
         </ul>
     </div>
 </dot-dropdown-component>
