+++ src/app/view/components/dot-toolbar/components/dot-toolbar-user/dot-toolbar-user.component.ts
@@ -2,10 +2,8 @@ import { Component, ViewChild, OnInit, Inject } from '@angular/core';
 import { DotDropdownComponent } from '@components/_common/dot-dropdown-component/dot-dropdown.component';
 import { IframeOverlayService } from '../../../_common/iframe/service/iframe-overlay.service';
 import { LoginService, Auth, LoggerService } from 'dotcms-js';
-import { DotMessageService } from '@services/dot-messages-service';
 import { LOCATION_TOKEN } from 'src/app/providers';
 import { DotNavigationService } from '@components/dot-navigation/services/dot-navigation.service';
-import { retryWhen, take, tap } from 'rxjs/operators';
 
 @Component({
     selector: 'dot-toolbar-user',
@@ -20,13 +18,8 @@ export class DotToolbarUserComponent implements OnInit {
     showLoginAs = false;
     showMyAccount = false;
 
-    i18nMessages: {
-        [key: string]: string;
-    } = {};
-
     constructor(
         @Inject(LOCATION_TOKEN) private location: Location,
-        private dotMessageService: DotMessageService,
         private loggerService: LoggerService,
         private loginService: LoginService,
         public iframeOverlayService: IframeOverlayService,
@@ -34,23 +27,6 @@ export class DotToolbarUserComponent implements OnInit {
     ) {}
 
     ngOnInit(): void {
-        this.dotMessageService
-            .getMessages(['my-account', 'login-as', 'Logout', 'logout-as'])
-            .pipe(
-                tap((res: {
-                    [key: string]: string;
-                }) => {
-                    if (!Object.keys(res).length) {
-                        throw new Error('No message keys');
-                    }
-                }),
-                retryWhen((errors) => errors),
-                take(1)
-            )
-            .subscribe((res) => {
-                this.i18nMessages = res;
-            });
-
         this.loginService.watchUser((auth: Auth) => {
             this.auth = auth;
         });
